Relevant BUGIDs:
Purpose of commit: bugfix

Commit summary:
---------------

2007-08-29  Thorsten Kukuk  <kukuk@thkukuk.de>

        * modules/pam_rhosts/pam_rhosts_auth.c: buflen needs to be size_t.
        (__icheckhost): Cast to int32_t to fix limited range error.

        * modules/pam_cracklib/pam_cracklib.c: Mark cracklib_dictpath
        as const.

+++ modules/pam_rhosts/pam_rhosts_auth.c
@@ -267,7 +267,7 @@ __icheckhost (pam_handle_t *pamh, struct _options *opts, u_int32_t raddr
 
 
     /* Try for raw ip address first. */
-    if (isdigit(*lhost) && (long)(laddr = inet_addr(lhost)) != -1)
+    if (isdigit(*lhost) && (int32_t)(laddr = inet_addr(lhost) != -1))
 	return (negate*(! (raddr ^ laddr)));
 
     /* Better be a hostname. */
@@ -350,7 +350,7 @@ __ivaliduser (pam_handle_t *pamh, struct _options *opts,
     int retval = 1;
 #ifdef HAVE_GETLINE
     char *buf=NULL;
-    int buflen=0;
+    size_t buflen=0;
 
     while (getline(&buf,&buflen,hostf) > 0) {
 #else
